+++ pkg/ttnpb/application.pb.validate.go
@@ -337,6 +337,89 @@ var _ interface {
 	ErrorName() string
 } = ApplicationsValidationError{}
 
+// ValidateFields checks the field values on IssueDevEUIResponse with the rules
+// defined in the proto definition for this message. If any rules are
+// violated, an error is returned.
+func (m *IssueDevEUIResponse) ValidateFields(paths ...string) error {
+	if m == nil {
+		return nil
+	}
+
+	if len(paths) == 0 {
+		paths = IssueDevEUIResponseFieldPathsNested
+	}
+
+	for name, subs := range _processPaths(append(paths[:0:0], paths...)) {
+		_ = subs
+		switch name {
+		case "dev_eui":
+			// no validation rules for DevEui
+		default:
+			return IssueDevEUIResponseValidationError{
+				field:  name,
+				reason: "invalid field path",
+			}
+		}
+	}
+	return nil
+}
+
+// IssueDevEUIResponseValidationError is the validation error returned by
+// IssueDevEUIResponse.ValidateFields if the designated constraints aren't met.
+type IssueDevEUIResponseValidationError struct {
+	field  string
+	reason string
+	cause  error
+	key    bool
+}
+
+// Field function returns field value.
+func (e IssueDevEUIResponseValidationError) Field() string { return e.field }
+
+// Reason function returns reason value.
+func (e IssueDevEUIResponseValidationError) Reason() string { return e.reason }
+
+// Cause function returns cause value.
+func (e IssueDevEUIResponseValidationError) Cause() error { return e.cause }
+
+// Key function returns key value.
+func (e IssueDevEUIResponseValidationError) Key() bool { return e.key }
+
+// ErrorName returns error name.
+func (e IssueDevEUIResponseValidationError) ErrorName() string {
+	return "IssueDevEUIResponseValidationError"
+}
+
+// Error satisfies the builtin error interface
+func (e IssueDevEUIResponseValidationError) Error() string {
+	cause := ""
+	if e.cause != nil {
+		cause = fmt.Sprintf(" | caused by: %v", e.cause)
+	}
+
+	key := ""
+	if e.key {
+		key = "key for "
+	}
+
+	return fmt.Sprintf(
+		"invalid %sIssueDevEUIResponse.%s: %s%s",
+		key,
+		e.field,
+		e.reason,
+		cause)
+}
+
+var _ error = IssueDevEUIResponseValidationError{}
+
+var _ interface {
+	Field() string
+	Reason() string
+	Key() bool
+	Cause() error
+	ErrorName() string
+} = IssueDevEUIResponseValidationError{}
+
 // ValidateFields checks the field values on GetApplicationRequest with the
 // rules defined in the proto definition for this message. If any rules are
 // violated, an error is returned.